+++ packages/terra-date-time-picker/tests/jest/DateTimePicker.test.jsx
@@ -1,33 +1,20 @@
 import React from 'react';
-import { IntlProvider } from 'react-intl';
+/* eslint-disable import/no-extraneous-dependencies */
+import { shallowWithIntl, mountWithIntl } from 'terra-enzyme-intl';
 import DateTimePicker from '../../lib/DateTimePicker';
-import messages from '../../translations/en-US.json';
-import dateInputMessages from '../../node_modules/terra-date-picker/translations/en-US.json';
-import timeInputMessages from '../../node_modules/terra-time-input/translations/en-US.json';
-
-const locale = 'en-US';
 
 it('should render a default date time picker', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DateTimePicker name="date-time-input" utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a default date time picker with custom date input attributes', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DateTimePicker name="date-time-input" utcOffset={0} dateInputAttributes={{ id: 'terra-date-input' }} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" utcOffset={0} dateInputAttributes={{ id: 'terra-date-input' }} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a default date time picker with custom time input attributes', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DateTimePicker name="date-time-input" utcOffset={0} timeInputAttributes={{ id: 'terra-time-input' }} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" utcOffset={0} timeInputAttributes={{ id: 'terra-time-input' }} />);
   expect(datePicker).toMatchSnapshot();
 });
 
@@ -36,96 +23,62 @@ it('should render a date time picker with filtered dates', () => {
     const day = date.day();
     return day !== 0 && day !== 6;
   };
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DateTimePicker filterDate={isWeekday} name="date-time-input" utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DateTimePicker filterDate={isWeekday} name="date-time-input" utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date time picker with disabled dates', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DateTimePicker name="date-time-input" excludeDates={['2017-04-01']} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" excludeDates={['2017-04-01']} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date time picker with included dates', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DateTimePicker name="date-time-input" includeDates={['2017-04-01']} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" includeDates={['2017-04-01']} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date time picker with min and max dates', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DateTimePicker name="date-time-input" maxDateTime="2017-04-01" minDateTime="2017-04-10" utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" maxDateTime="2017-04-01" minDateTime="2017-04-10" utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date time picker with releaseFocus', () => {
   const releaseFocus = () => {};
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DateTimePicker name="date-time-input" releaseFocus={releaseFocus} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" releaseFocus={releaseFocus} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date time picker with requestFocus', () => {
   const requestFocus = () => {};
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DateTimePicker name="date-time-input" requestFocus={requestFocus} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" requestFocus={requestFocus} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date time picker with onChange', () => {
   const onChange = () => {};
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DateTimePicker name="date-time-input" onChange={onChange} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" onChange={onChange} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date time picker with onChangeRaw', () => {
   const onChangeRaw = () => {};
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DateTimePicker name="date-time-input" onChangeRaw={onChangeRaw} utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" onChangeRaw={onChangeRaw} utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a date time picker with a default date and time', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DateTimePicker name="date-time-input" value="2017-04-01T10:30" utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" value="2017-04-01T10:30" utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('should render a disabled date time picker', () => {
-  const datePicker = shallow((
-    <IntlProvider locale={locale} messages={messages}>
-      <DateTimePicker name="date-time-input" disabled utcOffset={0} />
-    </IntlProvider>));
+  const datePicker = shallowWithIntl(<DateTimePicker name="date-time-input" disabled utcOffset={0} />);
   expect(datePicker).toMatchSnapshot();
 });
 
 it('Should not throw any errors while date value is outside of the Min, Max range and new time value is being entered ', () => {
-  const allMessages = Object.assign(messages, dateInputMessages, timeInputMessages);
-
-  const datePicker = mount((
-    <IntlProvider locale={locale} messages={allMessages}>
-      <DateTimePicker name="date-time-input" maxDateTime="2017-04-01T10:30" minDateTime="2017-04-10T10:30" />
-    </IntlProvider>
+  const datePicker = mountWithIntl((
+    <DateTimePicker name="date-time-input" maxDateTime="2017-04-01T10:30" minDateTime="2017-04-10T10:30" />
   ));
 
   const dateInput = datePicker.find({ name: 'terra-date-input', type: 'text' }).at(0);